fix: renaming core to sdk
