+++ package.json
@@ -2,9 +2,9 @@
   "name": "dc-visualization-sdk",
   "version": "0.0.1",
   "description": "",
-  "main": "dist/lib/dc-visualization-core.umd.js",
-  "module": "dist/lib/dc-visualization-core.es5.js",
-  "typings": "dist/types/dc-visualization-core.d.ts",
+  "main": "dist/lib/dc-visualization-sdk.umd.js",
+  "module": "dist/lib/dc-visualization-sdk.es5.js",
+  "typings": "dist/types/dc-visualization-sdk.d.ts",
   "scripts": {
     "prepare": "npm run build",
     "lint": "npm run prettier -- --check && npm run tslint",
@@ -12,7 +12,7 @@
     "tslint": "tslint  --project tsconfig.json -t codeFrame 'src/**/*.ts'",
     "prettier": "prettier \"src/**/*.ts\"",
     "prebuild": "rimraf dist",
-    "build": "tsc --module commonjs && rollup -c rollup.config.ts && typedoc --out docs src/dc-visualization-core.ts --media media/  --excludeInternal",
+    "build": "tsc --module commonjs && rollup -c rollup.config.ts && typedoc --out docs src/dc-visualization-sdk.ts --media media/  --excludeInternal",
     "build:watch": "tsc --module commonjs && rollup -c rollup.config.ts -w",
     "start": "rollup -c rollup.config.ts -w",
     "test": "jest --coverage",
