+++ rollup.config.ts
@@ -7,7 +7,7 @@ import json from 'rollup-plugin-json'
 
 const pkg = require('./package.json')
 
-const libraryName = 'dc-visualization-core'
+const libraryName = 'dc-visualization-sdk'
 
 export default {
   input: `src/${libraryName}.ts`,
